Remove clientTracking from WebSocket.Server
Previously, the `clientTracking` feature in WebSocket.Server was enabled. Unfortunately, in `socket.ts`, we remove all listeners from the `WebSocket` itself. This causes the `close` behavior removing the Socket from the `clients` list in `WebSocket.Server` not to function. Because the `terminate` functionality provided by `clientTracking` is never used with this project, the easiest solution is to remove it.